Moved file-patterns into the utils 🍭
